virtual: remove workspaces/kubeconfig subresource
