+++ test/e2e/virtual/workspaces/virtual_workspace_test.go
@@ -424,93 +424,6 @@ func testWorkspacesVirtualWorkspaces(t *testing.T, standalone bool) {
 				require.NoError(t, err, "did not see the workspace1 created in test org")
 			},
 		},
-		{
-			name: "create a workspace in personal virtual workspace and retrieve its kubeconfig",
-			virtualWorkspaceClientContexts: func(orgName logicalcluster.LogicalCluster) []clientInfo {
-				return []clientInfo{
-					{
-						// Use a user unique to the test to ensure isolation from other tests
-						Token:  "user-virtual-workspace-kubeconfig-token",
-						Prefix: path.Join(virtualoptions.DefaultRootPathPrefix, "workspaces", orgName.String(), "personal"),
-					},
-				}
-			},
-			work: func(ctx context.Context, t *testing.T, server runningServer) {
-				vwUser1Client := server.virtualKcpClients[0]
-
-				createOrgMemberRoleForGroup(t, ctx, server.kubeClusterClient, server.orgClusterName, "team-virtual-workspace-kubeconfig")
-
-				kcpServerKubeconfig, err := server.RunningServer.RawConfig()
-				require.NoError(t, err, "failed to get KCP Kubeconfig")
-
-				var workspace1 *tenancyv1beta1.Workspace
-				require.Eventually(t, func() bool {
-					// RBAC authz uses informers and needs a moment to understand the new roles. Hence, try until successful.
-					var err error
-					workspace1, err = vwUser1Client.TenancyV1beta1().Workspaces().Create(ctx, testData.workspace1.DeepCopy(), metav1.CreateOptions{})
-					if err != nil {
-						t.Logf("failed to create workspace1: %v", err)
-					}
-					return err == nil
-				}, wait.ForeverTestTimeout, time.Millisecond*100, "failed to create workspace1")
-
-				server.Artifact(t, func() (runtime.Object, error) {
-					return server.kcpClusterClient.Cluster(server.orgClusterName).TenancyV1alpha1().ClusterWorkspaces().Get(ctx, testData.workspace1.Name, metav1.GetOptions{})
-				})
-
-				require.Eventually(t, func() bool {
-					ws, err := server.kcpClusterClient.Cluster(server.orgClusterName).TenancyV1alpha1().ClusterWorkspaces().Get(ctx, workspace1.Name, metav1.GetOptions{})
-					if err != nil {
-						t.Logf("failed to retrieve workspace1: %v", err)
-						return false
-					}
-					return ws.Status.Phase == tenancyv1alpha1.ClusterWorkspacePhaseReady
-				}, wait.ForeverTestTimeout, time.Millisecond*100, " workspace1 did not become ready")
-
-				ws, err := server.kcpClusterClient.Cluster(server.orgClusterName).TenancyV1alpha1().ClusterWorkspaces().Get(ctx, workspace1.Name, metav1.GetOptions{})
-				require.NoError(t, err, "failed to get workspace1 as ClusterWorkspace")
-				workspaceURL := ws.Status.BaseURL
-
-				err = server.virtualWorkspaceExpectations[0](func(w *tenancyv1beta1.WorkspaceList) error {
-					if len(w.Items) != 1 || w.Items[0].Name != workspace1.Name {
-						return fmt.Errorf("expected only one workspace (%s), got %#v", workspace1.Name, w)
-					}
-					return nil
-				})
-				require.NoError(t, err, "did not see the workspace created in personal virtual workspace")
-
-				req := vwUser1Client.TenancyV1beta1().RESTClient().Get().Resource("workspaces").Name(workspace1.Name).SubResource("kubeconfig").Do(ctx)
-				require.Nil(t, req.Error(), "error retrieving the kubeconfig for workspace %s: %v", workspace1.Name, err)
-
-				kcpConfigCurrentContextName := kcpServerKubeconfig.CurrentContext
-				kcpConfigCurrentContext := kcpServerKubeconfig.Contexts[kcpConfigCurrentContextName]
-				require.NotNil(t, kcpConfigCurrentContext, "kcp Kubeconfig is invalid")
-
-				kcpConfigCurrentCluster := kcpServerKubeconfig.Clusters[kcpConfigCurrentContext.Cluster]
-				require.NotNil(t, kcpConfigCurrentCluster, "kcp Kubeconfig is invalid")
-
-				expectedKubeconfigCluster := kcpConfigCurrentCluster.DeepCopy()
-				expectedKubeconfigCluster.Server = workspaceURL
-				expectedKubeconfig := &clientcmdapi.Config{
-					CurrentContext: "personal/" + workspace1.Name,
-					Contexts: map[string]*clientcmdapi.Context{
-						"personal/" + workspace1.Name: {
-							Cluster: "personal/" + workspace1.Name,
-						},
-					},
-					Clusters: map[string]*clientcmdapi.Cluster{
-						"personal/" + workspace1.Name: expectedKubeconfigCluster,
-					},
-				}
-				expectedKubeconfigContent, err := clientcmd.Write(*expectedKubeconfig)
-				require.NoError(t, err, "error writing the content of the expected kubeconfig for workspace %s", workspace1.Name)
-
-				workspaceKubeconfigContent, err := req.Raw()
-				require.NoError(t, err, "error retrieving the content of the kubeconfig for workspace %s", workspace1.Name)
-
-				require.YAMLEq(t, string(expectedKubeconfigContent), string(workspaceKubeconfigContent))
-			},
-		},
 	}
 
 	var server framework.RunningServer
